+++ demisto_sdk/commands/common/tests/tools_test.py
@@ -9,12 +9,10 @@ import pytest
 import requests
 
 from demisto_sdk.commands.common import tools
-from demisto_sdk.commands.common.constants import (INTEGRATIONS_DIR,
-                                                   LAYOUTS_DIR, PACKS_DIR,
-                                                   PACKS_PACK_IGNORE_FILE_NAME,
-                                                   PLAYBOOKS_DIR, SCRIPTS_DIR,
-                                                   TEST_PLAYBOOKS_DIR,
-                                                   FileType)
+from demisto_sdk.commands.common.constants import (
+    DEFAULT_CONTENT_ITEM_TO_VERSION, INTEGRATIONS_DIR, LAYOUTS_DIR, PACKS_DIR,
+    PACKS_PACK_IGNORE_FILE_NAME, PLAYBOOKS_DIR, SCRIPTS_DIR,
+    TEST_PLAYBOOKS_DIR, FileType)
 from demisto_sdk.commands.common.legacy_git_tools import git_path
 from demisto_sdk.commands.common.tools import (
     LOG_COLORS, arg_to_list, compare_context_path_in_yml_and_readme,
@@ -28,7 +26,7 @@ from demisto_sdk.commands.common.tools import (
     get_release_note_entries, get_release_notes_file_path, get_ryaml,
     get_test_playbook_id, get_to_version, has_remote_configured,
     is_origin_content_repo, is_pack_path, is_uuid, retrieve_file_ending,
-    run_command_os, server_version_compare)
+    run_command_os, server_version_compare, to_kebab_case)
 from demisto_sdk.tests.constants_test import (IGNORED_PNG,
                                               INDICATORFIELD_EXTRA_FIELDS,
                                               SOURCE_FORMAT_INTEGRATION_COPY,
@@ -815,7 +813,7 @@ def test_get_to_version_no_to_version(repo):
     with ChangeCWD(repo.path):
         to_ver = get_to_version(integration.yml.path)
 
-        assert to_ver == '99.99.99'
+        assert to_ver == DEFAULT_CONTENT_ITEM_TO_VERSION
 
 
 def test_get_file_displayed_name__integration(repo):
@@ -1475,3 +1473,19 @@ def test_get_test_playbook_id():
     test_playbook_name, test_playbook_pack = get_test_playbook_id(test_playbook_id_set, test_name)
     assert test_playbook_name == 'HelloWorld-Test'
     assert test_playbook_pack == 'HelloWorld'
+
+
+KEBAB_CASES = [('Scan File', 'scan-file'),
+               ('Scan File-', 'scan-file'),
+               ('Scan.File', 'scan-file'),
+               ('*scan,file', 'scan-file'),
+               ('Scan     File', 'scan-file'),
+               ('Scan - File', 'scan-file'),
+               ('Scan-File', 'scan-file'),
+               ('Scan- File', 'scan-file'),
+               ('Scan -File', 'scan-file')]
+
+
+@pytest.mark.parametrize('input_str, output_str', KEBAB_CASES)
+def test_to_kebab_case(input_str, output_str):
+    assert to_kebab_case(input_str) == output_str